+++ src/main/java/csd/api/tables/PortfolioRepository.java
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Repository;
  * For the purpose of this exercise, CrudRepository would also be sufficient
  */
 @Repository
-public interface PortfolioRepository extends JpaRepository <Portfolio, Long> {
+public interface PortfolioRepository extends JpaRepository <Portfolio, Integer> {
 
 }
